barbershop/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
cryptography==36.0.2
dj-database-url==0.5.0
Django==4.0.3
django-cleanup==6.0.0
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
whitenoise==6.0.0
